Clean up integration tests. Consolidate logic. Allow the tests to be platform independent
