+++ android/services/mongodb-remote/src/androidTest/java/com/mongodb/stitch/android/services/mongodb/remote/internal/SyncMongoClientIntTests.kt
@@ -11,51 +11,104 @@ import com.mongodb.stitch.core.admin.authProviders.ProviderConfigs
 import com.mongodb.stitch.core.admin.services.ServiceConfigs
 import com.mongodb.stitch.core.admin.services.rules.RuleCreator
 import com.mongodb.stitch.core.auth.providers.anonymous.AnonymousCredential
-import com.mongodb.stitch.core.internal.common.Callback
-import com.mongodb.stitch.core.internal.common.OperationResult
+import com.mongodb.stitch.core.services.mongodb.remote.RemoteDeleteResult
+import com.mongodb.stitch.core.services.mongodb.remote.RemoteInsertManyResult
+import com.mongodb.stitch.core.services.mongodb.remote.RemoteInsertOneResult
+import com.mongodb.stitch.core.services.mongodb.remote.RemoteUpdateResult
+import com.mongodb.stitch.core.services.mongodb.remote.sync.ChangeEventListener
 import com.mongodb.stitch.core.services.mongodb.remote.sync.ConflictHandler
-import com.mongodb.stitch.core.services.mongodb.remote.sync.internal.ChangeEvent
-import com.mongodb.stitch.core.services.mongodb.remote.sync.DefaultSyncConflictResolvers
 import com.mongodb.stitch.core.services.mongodb.remote.sync.ErrorListener
-import org.bson.BsonDocument
-import org.bson.BsonObjectId
+import com.mongodb.stitch.core.services.mongodb.remote.sync.internal.DataSynchronizer
+import com.mongodb.stitch.core.testutils.BaseStitchIntTest
+import com.mongodb.stitch.core.testutils.sync.CoreRemoteMethods
+import com.mongodb.stitch.core.testutils.sync.CoreSyncMethods
+import com.mongodb.stitch.core.testutils.sync.SyncIntHarness
+import com.mongodb.stitch.core.testutils.sync.SyncIntTestRunner
 import org.bson.BsonValue
 import org.bson.Document
+import org.bson.conversions.Bson
 import org.bson.types.ObjectId
 import org.junit.After
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertNotEquals
-import org.junit.Assert.assertNotNull
-import org.junit.Assert.assertNull
-import org.junit.Assert.assertTrue
-import org.junit.Assert.fail
+import org.junit.Assert
 import org.junit.Assume
 import org.junit.Before
 import org.junit.Test
-import java.lang.Exception
-import java.util.UUID
-import java.util.concurrent.Semaphore
-import java.util.concurrent.TimeUnit
-import java.util.concurrent.atomic.AtomicInteger
 
-class SyncMongoClientIntTests : BaseStitchAndroidIntTest() {
+class SyncMongoClientIntTests : BaseStitchAndroidIntTest(), SyncIntTestRunner {
+    private class RemoteMethods(private val remoteMongoCollection: RemoteMongoCollection<Document>): CoreRemoteMethods {
+        override fun insertOne(document: Document): RemoteInsertOneResult {
+            return Tasks.await(remoteMongoCollection.insertOne(document))
+        }
+        override fun insertMany(documents: List<Document>): RemoteInsertManyResult {
+            return Tasks.await(remoteMongoCollection.insertMany(documents))
+        }
 
-    private val mongodbUriProp = "test.stitch.mongodbURI"
+        override fun find(filter: Document): Iterable<Document?> {
+            return Tasks.await(remoteMongoCollection.find(filter).into(mutableListOf<Document>()))
+        }
 
-    private var remoteMongoClientOpt: RemoteMongoClient? = null
-    private val remoteMongoClient: RemoteMongoClient
-        get() = remoteMongoClientOpt!!
-    private var mongoClientOpt: RemoteMongoClient? = null
-    private val mongoClient: RemoteMongoClient
-        get() = mongoClientOpt!!
+        override fun updateOne(filter: Document, updateDocument: Document): RemoteUpdateResult {
+            return Tasks.await(remoteMongoCollection.updateOne(filter, updateDocument))
+        }
+
+        override fun deleteOne(filter: Bson): RemoteDeleteResult {
+            return Tasks.await(remoteMongoCollection.deleteOne(filter))
+        }
+    }
+
+    private class SyncMethods(private val sync: Sync<Document>): CoreSyncMethods {
+        override fun configure(conflictResolver: ConflictHandler<Document?>,
+                               changeEventListener: ChangeEventListener<Document>?,
+                               errorListener: ErrorListener?) {
+            sync.configure(conflictResolver, changeEventListener, errorListener)
+        }
+
+        override fun syncOne(id: BsonValue) {
+            sync.syncOne(id)
+        }
+
+        override fun insertOneAndSync(document: Document): RemoteInsertOneResult {
+            return Tasks.await(sync.insertOneAndSync(document))
+        }
+
+        override fun findOneById(id: BsonValue): Document? {
+            return Tasks.await(sync.findOneById(id))
+        }
+
+        override fun updateOneById(documentId: BsonValue, update: Bson): RemoteUpdateResult {
+            return Tasks.await(sync.updateOneById(documentId, update))
+        }
+
+        override fun deleteOneById(documentId: BsonValue): RemoteDeleteResult {
+            return Tasks.await(sync.deleteOneById(documentId))
+        }
+
+        override fun desyncOne(id: BsonValue) {
+            sync.desyncOne(id)
+        }
+
+        override fun getSyncedIds(): Set<BsonValue> {
+            return sync.syncedIds
+        }
+
+        override fun find(filter: Bson): Iterable<Document?> {
+            return Tasks.await(sync.find(filter).into(mutableListOf<Document>()))
+        }
+    }
+
     private var dbName = ObjectId().toHexString()
     private var collName = ObjectId().toHexString()
-    private var namespace = MongoNamespace(dbName, collName)
+    override var namespace = MongoNamespace(dbName, collName)
+    override val dataSynchronizer: DataSynchronizer
+        get() = (mongoClient as RemoteMongoClientImpl).dataSynchronizer
+    override val testNetworkMonitor: BaseStitchIntTest.TestNetworkMonitor
+        get() = BaseStitchAndroidIntTest.testNetworkMonitor
 
-    private fun getMongoDbUri(): String {
-        return InstrumentationRegistry.getArguments().getString(mongodbUriProp, "mongodb://localhost:26000")
-    }
+    private val mongodbUriProp = "test.stitch.mongodbURI"
+    private lateinit var remoteMongoClient: RemoteMongoClient
+    private lateinit var mongoClient: RemoteMongoClient
+
+    private val harness = SyncIntHarness(this)
 
     @Before
     override fun setup() {
@@ -68,15 +121,15 @@ class SyncMongoClientIntTests : BaseStitchAndroidIntTest() {
         addProvider(app.second, ProviderConfigs.Anon)
         addProvider(app2.second, ProviderConfigs.Anon)
         val svc = addService(
-                app.second,
-                "mongodb",
-                "mongodb1",
-                ServiceConfigs.Mongo(getMongoDbUri()))
+            app.second,
+            "mongodb",
+            "mongodb1",
+            ServiceConfigs.Mongo(getMongoDbUri()))
         val svc2 = addService(
-                app2.second,
-                "mongodb",
-                "mongodb1",
-                ServiceConfigs.Mongo(getMongoDbUri()))
+            app2.second,
+            "mongodb",
+            "mongodb1",
+            ServiceConfigs.Mongo(getMongoDbUri()))
 
         val rule = Document()
         rule["read"] = Document()
@@ -91,12 +144,12 @@ class SyncMongoClientIntTests : BaseStitchAndroidIntTest() {
         addRule(svc2.second, RuleCreator.MongoDb("$dbName.$collName", rule))
 
         val client = getAppClient(app.first)
-        Tasks.await(client.auth.loginWithCredential(AnonymousCredential()))
-        mongoClientOpt = client.getServiceClient(RemoteMongoClient.factory, "mongodb1")
+        client.auth.loginWithCredential(AnonymousCredential())
+        mongoClient = client.getServiceClient(RemoteMongoClient.factory, "mongodb1")
         (mongoClient as RemoteMongoClientImpl).dataSynchronizer.stop()
         (mongoClient as RemoteMongoClientImpl).dataSynchronizer.disableSyncThread()
-        remoteMongoClientOpt = client.getServiceClient(RemoteMongoClient.factory, "mongodb1")
-        goOnline()
+        remoteMongoClient = client.getServiceClient(RemoteMongoClient.factory, "mongodb1")
+        BaseStitchAndroidIntTest.testNetworkMonitor.connectedState = true
     }
 
     @After
@@ -105,1200 +158,146 @@ class SyncMongoClientIntTests : BaseStitchAndroidIntTest() {
         super.teardown()
     }
 
-    private fun getTestSync(): Sync<Document> {
-        val db = mongoClient.getDatabase(dbName)
-        assertEquals(dbName, db.name)
-        val coll = db.getCollection(collName)
-        assertEquals(MongoNamespace(dbName, collName), coll.namespace)
-        return coll.sync()
-    }
-
-    private fun getTestCollRemote(): RemoteMongoCollection<Document> {
+    override fun remoteMethods(): CoreRemoteMethods {
         val db = remoteMongoClient.getDatabase(dbName)
-        assertEquals(dbName, db.name)
+        Assert.assertEquals(dbName, db.name)
         val coll = db.getCollection(collName)
-        assertEquals(MongoNamespace(dbName, collName), coll.namespace)
-        return coll
+        Assert.assertEquals(MongoNamespace(dbName, collName), coll.namespace)
+        return RemoteMethods(coll)
+    }
+
+    override fun syncMethods(): CoreSyncMethods {
+        val db = mongoClient.getDatabase(dbName)
+        Assert.assertEquals(dbName, db.name)
+        val coll = db.getCollection(collName)
+        Assert.assertEquals(MongoNamespace(dbName, collName), coll.namespace)
+        return SyncMethods(coll.sync())
     }
 
     @Test
-    fun testSync() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val doc1 = Document("hello", "world")
-            val doc2 = Document("hello", "friend")
-            doc2["proj"] = "field"
-            Tasks.await(remoteColl.insertMany(listOf(doc1, doc2)))
-
-            // get the document
-            val doc = Tasks.await(remoteColl.find(doc1).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            // start watching it and always set the value to hello world in a conflict
-            coll.configure({ id: BsonValue, localEvent: ChangeEvent<Document>, remoteEvent: ChangeEvent<Document> ->
-                if (id.equals(doc1Id)) {
-                    val merged = localEvent.fullDocument.getInteger("foo") +
-                            remoteEvent.fullDocument.getInteger("foo")
-                    val newDocument = Document(HashMap<String, Any>(remoteEvent.fullDocument))
-                    newDocument["foo"] = merged
-                    newDocument
-                } else {
-                    Document("hello", "world")
-                }
-            }, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-
-            // 1. updating a document remotely should not be reflected until coming back online.
-            goOffline()
-            val doc1Update = Document("\$inc", Document("foo", 1))
-            val result = Tasks.await(remoteColl.updateOne(
-                    doc1Filter,
-                    doc1Update))
-            assertEquals(1, result.matchedCount)
-            streamAndSync()
-            assertEquals(doc, Tasks.await(coll.findOneById(doc1Id)))
-            goOnline()
-            streamAndSync()
-            val expectedDocument = Document(doc)
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, Tasks.await(coll.findOneById(doc1Id)))
-
-            // 2. insertOneAndSync should work offline and then sync the document when online.
-            goOffline()
-            val doc3 = Document("so", "syncy")
-            val insResult = Tasks.await(coll.insertOneAndSync(doc3))
-            assertEquals(doc3, withoutSyncVersion(Tasks.await(coll.findOneById(insResult.insertedId))!!))
-            streamAndSync()
-            assertNull(Tasks.await(remoteColl.find(Document("_id", doc3["_id"])).first()))
-            goOnline()
-            streamAndSync()
-            assertEquals(doc3, withoutSyncVersion(Tasks.await(remoteColl.find(Document("_id", doc3["_id"])).first())!!))
-
-            // 3. updating a document locally that has been updated remotely should invoke the conflict
-            // resolver.
-            val sem = watchForEvents(this.namespace)
-            val result2 = Tasks.await(remoteColl.updateOne(
-                    doc1Filter,
-                    withNewSyncVersionSet(doc1Update)))
-            sem.acquire()
-            assertEquals(1, result2.matchedCount)
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            val result3 = Tasks.await(coll.updateOneById(
-                    doc1Id,
-                    doc1Update))
-            assertEquals(1, result3.matchedCount)
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            // first pass will invoke the conflict handler and update locally but not remotely yet
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            expectedDocument["foo"] = 4
-            expectedDocument.remove("fooOps")
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            // second pass will update with the ack'd version id
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-        }
+    override fun testSync() {
+        harness.testSync()
     }
 
     @Test
-    fun testUpdateConflicts() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure({ _: BsonValue, localEvent: ChangeEvent<Document>, remoteEvent: ChangeEvent<Document> ->
-                val merged = Document(localEvent.fullDocument)
-                remoteEvent.fullDocument.forEach {
-                    if (localEvent.fullDocument.containsKey(it.key)) {
-                        return@forEach
-                    }
-                    merged[it.key] = it.value
-                }
-                merged
-            }, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-
-            // Update remote
-            val remoteUpdate = withNewSyncVersionSet(Document("\$set", Document("remote", "update")))
-            val sem = watchForEvents(this.namespace)
-            var result = Tasks.await(remoteColl.updateOne(doc1Filter, remoteUpdate))
-            sem.acquire()
-            assertEquals(1, result.matchedCount)
-            val expectedRemoteDocument = Document(doc)
-            expectedRemoteDocument["remote"] = "update"
-            assertEquals(expectedRemoteDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-
-            // Update local
-            val localUpdate = Document("\$set", Document("local", "updateWow"))
-            result = Tasks.await(coll.updateOneById(doc1Id, localUpdate))
-            assertEquals(1, result.matchedCount)
-            val expectedLocalDocument = Document(doc)
-            expectedLocalDocument["local"] = "updateWow"
-            assertEquals(expectedLocalDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            // first pass will invoke the conflict handler and update locally but not remotely yet
-            streamAndSync()
-            assertEquals(expectedRemoteDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            expectedLocalDocument["remote"] = "update"
-            assertEquals(expectedLocalDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            // second pass will update with the ack'd version id
-            streamAndSync()
-            assertEquals(expectedLocalDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            assertEquals(expectedLocalDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-        }
+    override fun testUpdateConflicts() {
+        harness.testUpdateConflicts()
     }
 
     @Test
-    fun testUpdateRemoteWins() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            docToInsert["foo"] = 1
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure(DefaultSyncConflictResolvers.remoteWins(), null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-
-            val expectedDocument = Document(doc)
-            val sem = watchForEvents(this.namespace)
-            var result = Tasks.await(remoteColl.updateOne(doc1Filter, withNewSyncVersionSet(Document("\$inc", Document("foo", 2)))))
-            sem.acquire()
-            assertEquals(1, result.matchedCount)
-            expectedDocument["foo"] = 3
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            result = Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1))))
-            assertEquals(1, result.matchedCount)
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            streamAndSync()
-            expectedDocument["foo"] = 3
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-        }
+    override fun testUpdateRemoteWins() {
+        harness.testUpdateRemoteWins()
     }
 
     @Test
-    fun testUpdateLocalWins() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            docToInsert["foo"] = 1
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-
-            val expectedDocument = Document(doc)
-            val sem = watchForEvents(this.namespace)
-            var result = Tasks.await(remoteColl.updateOne(doc1Filter, withNewSyncVersionSet(Document("\$inc", Document("foo", 2)))))
-            sem.acquire()
-            assertEquals(1, result.matchedCount)
-            expectedDocument["foo"] = 3
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            result = Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1))))
-            assertEquals(1, result.matchedCount)
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            streamAndSync()
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-        }
+    override fun testUpdateLocalWins() {
+        harness.testUpdateLocalWins()
     }
 
     @Test
-    fun testDeleteOneByIdNoConflict() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure(failingConflictHandler, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-
-            goOffline()
-            val result = Tasks.await(coll.deleteOneById(doc1Id))
-            assertEquals(1, result.deletedCount)
-
-            val expectedDocument = withoutSyncVersion(Document(doc))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            goOnline()
-            streamAndSync()
-            assertNull(Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testDeleteOneByIdNoConflict() {
+        harness.testDeleteOneByIdNoConflict()
     }
 
     @Test
-    fun testDeleteOneByIdConflict() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                Document("well", "shoot")
-            }, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-
-            val doc1Update = Document("\$inc", Document("foo", 1))
-            assertEquals(1, Tasks.await(remoteColl.updateOne(
-                    doc1Filter,
-                    withNewSyncVersionSet(doc1Update))).matchedCount)
-
-            goOffline()
-            val result = Tasks.await(coll.deleteOneById(doc1Id))
-            assertEquals(1, result.deletedCount)
-
-            val expectedDocument = Document(doc)
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            goOnline()
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            expectedDocument.remove("hello")
-            expectedDocument.remove("foo")
-            expectedDocument["well"] = "shoot"
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-        }
+    override fun testDeleteOneByIdConflict() {
+        harness.testDeleteOneByIdConflict()
     }
 
     @Test
-    fun testInsertThenUpdateThenSync() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-
-            coll.configure(failingConflictHandler, null, null)
-            val insertResult = Tasks.await(coll.insertOneAndSync(docToInsert))
-
-            val doc = Tasks.await(coll.findOneById(insertResult.insertedId))!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            val doc1Update = Document("\$inc", Document("foo", 1))
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, doc1Update)).matchedCount)
-
-            val expectedDocument = withoutSyncVersion(Document(doc))
-            expectedDocument["foo"] = 1
-            assertNull(Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            goOnline()
-            streamAndSync()
-
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-        }
+    override fun testInsertThenUpdateThenSync() {
+        harness.testInsertThenUpdateThenSync()
     }
 
     @Test
-    fun testInsertThenSyncUpdateThenUpdate() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-
-            coll.configure(failingConflictHandler, null, null)
-            val insertResult = Tasks.await(coll.insertOneAndSync(docToInsert))
-
-            val doc = Tasks.await(coll.findOneById(insertResult.insertedId))!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            goOnline()
-            streamAndSync()
-            val expectedDocument = withoutSyncVersion(Document(doc))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            val doc1Update = Document("\$inc", Document("foo", 1))
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, doc1Update)).matchedCount)
-
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-        }
+    override fun testInsertThenSyncUpdateThenUpdate() {
+        harness.testInsertThenSyncUpdateThenUpdate()
     }
 
     @Test
-    fun testInsertThenSyncThenRemoveThenInsertThenUpdate() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            coll.configure(failingConflictHandler, null, null)
-            val insertResult = Tasks.await(coll.insertOneAndSync(docToInsert))
-            streamAndSync()
-
-            val doc = Tasks.await(coll.findOneById(insertResult.insertedId))!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-            val expectedDocument = withoutSyncVersion(Document(doc))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            assertEquals(1, Tasks.await(coll.deleteOneById(doc1Id)).deletedCount)
-            Tasks.await(coll.insertOneAndSync(doc))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            val doc1Update = Document("\$inc", Document("foo", 1))
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, doc1Update)).matchedCount)
-
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-        }
+    override fun testInsertThenSyncThenRemoveThenInsertThenUpdate() {
+        harness.testInsertThenSyncThenRemoveThenInsertThenUpdate()
     }
 
     @Test
-    fun testRemoteDeletesLocalNoConflict() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure(failingConflictHandler, null, null)
-            coll.syncOne(doc1Id)
-
-            streamAndSync()
-
-            assertEquals(coll.syncedIds.size, 1)
-
-            val sem = watchForEvents(this.namespace)
-            Tasks.await(remoteColl.deleteOne(doc1Filter))
-            sem.acquire()
-
-            streamAndSync()
-
-            assertNull(Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            // This should not re-sync the document
-            streamAndSync()
-            Tasks.await(remoteColl.insertOne(doc))
-            streamAndSync()
-
-            assertEquals(doc, Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testRemoteDeletesLocalNoConflict() {
+        harness.testRemoteDeletesLocalNoConflict()
     }
 
     @Test
-    fun testRemoteDeletesLocalConflict() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                Document("hello", "world")
-            }, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-            assertEquals(doc, Tasks.await(coll.findOneById(doc1Id)))
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            goOffline()
-            Tasks.await(remoteColl.deleteOne(doc1Filter))
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1)))).matchedCount)
-
-            goOnline()
-            streamAndSync()
-            assertNull(Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            streamAndSync()
-            assertNotNull(Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testRemoteDeletesLocalConflict() {
+        harness.testRemoteDeletesLocalConflict()
     }
 
     @Test
-    fun testRemoteInsertsLocalUpdates() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                Document("hello", "again")
-            }, null, null)
-            coll.syncOne(doc1Id)
-
-            streamAndSync()
-
-            assertEquals(doc, Tasks.await(coll.findOneById(doc1Id)))
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            val wait = watchForEvents(this.namespace, 2)
-            Tasks.await(remoteColl.deleteOne(doc1Filter))
-            Tasks.await(remoteColl.insertOne(withNewSyncVersion(doc)))
-            wait.acquire()
-
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1)))).matchedCount)
-
-            streamAndSync()
-
-            assertEquals(doc, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            val expectedDocument = Document("_id", doc1Id.value)
-            expectedDocument["hello"] = "again"
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-        }
+    override fun testRemoteInsertsLocalUpdates() {
+        harness.testRemoteInsertsLocalUpdates()
     }
 
     @Test
-    fun testRemoteInsertsWithVersionLocalUpdates() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(withNewSyncVersion(docToInsert)))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure(failingConflictHandler, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-            assertEquals(doc, Tasks.await(coll.findOneById(doc1Id)))
-
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1)))).matchedCount)
-
-            streamAndSync()
-            val expectedDocument = Document(withoutSyncVersion(doc))
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-        }
+    override fun testRemoteInsertsWithVersionLocalUpdates() {
+        harness.testRemoteInsertsWithVersionLocalUpdates()
     }
 
     @Test
-    fun testResolveConflictWithDelete() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(withNewSyncVersion(docToInsert)))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                null
-            }, null, null)
-            coll.syncOne(doc1Id)
-            streamAndSync()
-            assertEquals(doc, Tasks.await(coll.findOneById(doc1Id)))
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            val sem = watchForEvents(this.namespace)
-            assertEquals(1, Tasks.await(remoteColl.updateOne(doc1Filter, withNewSyncVersionSet(Document("\$inc", Document("foo", 1))))).matchedCount)
-            sem.acquire()
-
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1)))).matchedCount)
-
-            streamAndSync()
-            val expectedDocument = Document(withoutSyncVersion(doc))
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-
-            goOffline()
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            goOnline()
-            streamAndSync()
-            assertNull(Tasks.await(remoteColl.find(doc1Filter).first()))
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testResolveConflictWithDelete() {
+        harness.testResolveConflictWithDelete()
     }
 
     @Test
-    fun testTurnDeviceOffAndOn() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-            docToInsert["foo"] = 1
-            Tasks.await(remoteColl.insertOne(docToInsert))
-
-            val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            powerCycleDevice()
-
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-            coll.syncOne(doc1Id)
-
-            powerCycleDevice()
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-            streamAndSync()
-
-            val expectedDocument = Document(doc)
-            var result = Tasks.await(remoteColl.updateOne(doc1Filter, withNewSyncVersionSet(Document("\$inc", Document("foo", 2)))))
-            assertEquals(1, result.matchedCount)
-            expectedDocument["foo"] = 3
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-            powerCycleDevice()
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-
-            result = Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("foo", 1))))
-            assertEquals(1, result.matchedCount)
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            powerCycleDevice()
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-
-            streamAndSync() // does nothing with no conflict handler
-
-            assertEquals(1, coll.syncedIds.size)
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-            streamAndSync() // resolves the conflict
-
-            expectedDocument["foo"] = 2
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            powerCycleDevice()
-            coll.configure(DefaultSyncConflictResolvers.localWins(), null, null)
-
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-        }
+    override fun testTurnDeviceOffAndOn() {
+        harness.testTurnDeviceOffAndOn()
     }
 
     @Test
-    fun testDesync() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val docToInsert = Document("hello", "world")
-            coll.configure(failingConflictHandler, null, null)
-            val doc1Id = Tasks.await(coll.insertOneAndSync(docToInsert)).insertedId
-
-            assertEquals(docToInsert, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            coll.desyncOne(doc1Id)
-            streamAndSync()
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testDesync() {
+        harness.testDesync()
     }
 
     @Test
-    fun testInsertInsertConflict() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("_id", "hello")
-
-            Tasks.await(remoteColl.insertOne(docToInsert))
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                Document("friend", "welcome")
-            }, null, null)
-            val doc1Id = Tasks.await(coll.insertOneAndSync(docToInsert)).insertedId
-
-            val doc1Filter = Document("_id", doc1Id)
-
-            streamAndSync()
-            val expectedDocument = Document(docToInsert)
-            expectedDocument["friend"] = "welcome"
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            assertEquals(docToInsert, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-
-            streamAndSync()
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(remoteColl.find(doc1Filter).first())!!))
-        }
+    override fun testInsertInsertConflict() {
+        harness.testInsertInsertConflict()
     }
 
     @Test
-    fun testFrozenDocumentConfig() {
-        testSyncInBothDirections {
-            val testSync = getTestSync()
-            val remoteColl = getTestCollRemote()
-            var errorEmitted = false
-
-            var conflictCounter = 0
-
-            testSync.configure(
-                    { _: BsonValue, _: ChangeEvent<Document>, remoteEvent: ChangeEvent<Document> ->
-                        if (conflictCounter == 0) {
-                            conflictCounter++
-                            errorEmitted = true
-                            throw Exception("ouch")
-                        }
-                        remoteEvent.fullDocument
-                    },
-                    { _: BsonValue, _: ChangeEvent<Document> ->
-                    }, { _, _ ->
-            })
-
-            // insert an initial doc
-            val testDoc = Document("hello", "world")
-            val result = Tasks.await(testSync.insertOneAndSync(testDoc))
-
-            // do a sync pass, synchronizing the doc
-            streamAndSync()
-
-            assertNotNull(Tasks.await(remoteColl.find(Document("_id", testDoc.get("_id"))).first()))
-
-            // update the doc
-            val expectedDoc = Document("hello", "computer")
-            Tasks.await(testSync.updateOneById(result.insertedId, Document("\$set", expectedDoc)))
-
-            // create a conflict
-            var sem = watchForEvents(namespace)
-            Tasks.await(remoteColl.updateOne(Document("_id", result.insertedId), withNewSyncVersionSet(Document("\$inc", Document("foo", 2)))))
-            sem.acquire()
-
-            // do a sync pass, and throw an error during the conflict resolver
-            // freezing the document
-            streamAndSync()
-            assertTrue(errorEmitted)
-
-            // update the doc remotely
-            val nextDoc = Document("hello", "friend")
-
-            sem = watchForEvents(namespace)
-            Tasks.await(remoteColl.updateOne(Document("_id", result.insertedId), nextDoc))
-            sem.acquire()
-            streamAndSync()
-
-            // it should not have updated the local doc, as the local doc should be frozen
-            assertEquals(
-                    withoutId(expectedDoc),
-                    withoutSyncVersion(withoutId(Tasks.await(testSync.find(Document("_id", result.insertedId)).first())!!)))
-
-            // update the local doc. this should unfreeze the config
-            Tasks.await(testSync.updateOneById(result.insertedId, Document("\$set", Document("no", "op"))))
-
-            streamAndSync()
-
-            // this should still be the remote doc since remote wins
-            assertEquals(
-                    withoutId(nextDoc),
-                    withoutSyncVersion(withoutId(Tasks.await(testSync.find(Document("_id", result.insertedId)).first())!!)))
-
-            // update the doc remotely
-            val lastDoc = Document("good night", "computer")
-
-            sem = watchForEvents(namespace)
-            Tasks.await(remoteColl.updateOne(
-                    Document("_id", result.insertedId),
-                    withNewSyncVersion(lastDoc)
-            ))
-            sem.acquire()
-
-            // now that we're sync'd and unfrozen, it should be reflected locally
-            // TODO: STITCH-1958 Possible race condition here for update listening
-            streamAndSync()
-
-            assertEquals(
-                    withoutId(lastDoc),
-                    withoutSyncVersion(
-                            withoutId(Tasks.await(testSync.find(Document("_id", result.insertedId)).first())!!)))
-        }
+    override fun testFrozenDocumentConfig() {
+        harness.testFrozenDocumentConfig()
     }
 
     @Test
-    fun testConfigure() {
-        val testSync = getTestSync()
-        val remoteColl = getTestCollRemote()
-
-        val docToInsert = Document("hello", "world")
-        val insertedId = Tasks.await(testSync.insertOneAndSync(docToInsert)).insertedId
-
-        var hasConflictHandlerBeenInvoked = false
-        var hasChangeEventListenerBeenInvoked = false
-
-        testSync.configure(
-            { _: BsonValue, _: ChangeEvent<Document>, remoteEvent: ChangeEvent<Document> ->
-                hasConflictHandlerBeenInvoked = true
-                assertEquals(remoteEvent.fullDocument["fly"], "away")
-                remoteEvent.fullDocument
-            },
-            { _: BsonValue, _: ChangeEvent<Document> ->
-                hasChangeEventListenerBeenInvoked = true
-            },
-            { _, _ -> }
-        )
-
-        val sem = watchForEvents(namespace)
-        Tasks.await(remoteColl.insertOne(Document("_id", insertedId).append("fly", "away")))
-        sem.acquire()
-
-        streamAndSync()
-
-        assertTrue(hasConflictHandlerBeenInvoked)
-        assertTrue(hasChangeEventListenerBeenInvoked)
+    override fun testConfigure() {
+        harness.testConfigure()
     }
 
     @Test
-    fun testSyncVersioningScheme() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val docToInsert = Document("hello", "world")
-
-            coll.configure(failingConflictHandler, null, null)
-            val insertResult = Tasks.await(coll.insertOneAndSync(docToInsert))
-
-            val doc = Tasks.await(coll.findOneById(insertResult.insertedId))!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            goOnline()
-            streamAndSync()
-            val expectedDocument = Document(doc)
-
-            // the remote document after an initial insert should have a fresh instance ID, and a
-            // version counter of 0
-            val firstRemoteDoc = Tasks.await(remoteColl.find(doc1Filter).first())!!
-            assertEquals(expectedDocument, withoutSyncVersion(firstRemoteDoc))
-
-            assertEquals(0, versionCounterOf(firstRemoteDoc))
-
-            assertEquals(expectedDocument, Tasks.await(coll.findOneById(doc1Id))!!)
-
-            // the remote document after a local update, but before a sync pass, should have the
-            // same version as the original document, and be equivalent to the unupdated document
-            val doc1Update = Document("\$inc", Document("foo", 1))
-            assertEquals(1, Tasks.await(coll.updateOneById(doc1Id, doc1Update)).matchedCount)
-
-            val secondRemoteDocBeforeSyncPass = Tasks.await(remoteColl.find(doc1Filter).first())!!
-            assertEquals(expectedDocument, withoutSyncVersion(secondRemoteDocBeforeSyncPass))
-            assertEquals(versionOf(firstRemoteDoc), versionOf(secondRemoteDocBeforeSyncPass))
-
-            expectedDocument["foo"] = 1
-            assertEquals(expectedDocument, Tasks.await(coll.findOneById(doc1Id))!!)
-
-            // the remote document after a local update, and after a sync pass, should have a new
-            // version with the same instance ID as the original document, a version counter
-            // incremented by 1, and be equivalent to the updated document.
-            streamAndSync()
-            val secondRemoteDoc = Tasks.await(remoteColl.find(doc1Filter).first())!!
-            assertEquals(expectedDocument, withoutSyncVersion(secondRemoteDoc))
-            assertEquals(instanceIdOf(firstRemoteDoc), instanceIdOf(secondRemoteDoc))
-            assertEquals(1, versionCounterOf(secondRemoteDoc))
-
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))!!))
-
-            // the remote document after a local delete and local insert, but before a sync pass,
-            // should have the same version as the previous document
-            assertEquals(1, Tasks.await(coll.deleteOneById(doc1Id)).deletedCount)
-            Tasks.await(coll.insertOneAndSync(doc))
-
-            val thirdRemoteDocBeforeSyncPass = Tasks.await(remoteColl.find(doc1Filter).first())!!
-            assertEquals(expectedDocument, withoutSyncVersion(thirdRemoteDocBeforeSyncPass))
-
-            expectedDocument.remove("foo")
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))))
-
-            // the remote document after a local delete and local insert, and after a sync pass,
-            // should have the same instance ID as before and a version count, since the change
-            // events are coalesced into a single update event
-            streamAndSync()
-
-            val thirdRemoteDoc = Tasks.await(remoteColl.find(doc1Filter).first())!!
-            assertEquals(expectedDocument, withoutSyncVersion(thirdRemoteDoc))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))))
-
-            assertEquals(instanceIdOf(secondRemoteDoc), instanceIdOf(thirdRemoteDoc))
-            assertEquals(2, versionCounterOf(thirdRemoteDoc))
-
-            // the remote document after a local delete, a sync pass, a local insert, and after
-            // another sync pass should have a new instance ID, with a version counter of zero,
-            // since the change events are not coalesced
-            assertEquals(1, Tasks.await(coll.deleteOneById(doc1Id)).deletedCount)
-            streamAndSync()
-            Tasks.await(coll.insertOneAndSync(doc))
-            streamAndSync()
-
-            val fourthRemoteDoc = Tasks.await(remoteColl.find(doc1Filter).first())!!
-            assertEquals(expectedDocument, withoutSyncVersion(thirdRemoteDoc))
-            assertEquals(expectedDocument, withoutSyncVersion(Tasks.await(coll.findOneById(doc1Id))))
-
-            assertNotEquals(instanceIdOf(secondRemoteDoc), instanceIdOf(fourthRemoteDoc))
-            assertEquals(0, versionCounterOf(fourthRemoteDoc))
-        }
+    override fun testSyncVersioningScheme() {
+        harness.testSyncVersioningScheme()
     }
 
     @Test
-    fun testUnsupportedSpvFails() {
-        val coll = getTestSync()
-
-        val remoteColl = getTestCollRemote()
-
-        val docToInsert = withNewUnsupportedSyncVersion(Document("hello", "world"))
-
-        val errorEmittedSem = Semaphore(0)
-        coll.configure(
-                failingConflictHandler,
-                null,
-                ErrorListener { documentId, error -> errorEmittedSem.release() })
-
-        Tasks.await(remoteColl.insertOne(docToInsert))
-
-        val doc = Tasks.await(remoteColl.find(docToInsert).first())!!
-        val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-        coll.syncOne(doc1Id)
-
-        assertTrue(coll.syncedIds.contains(doc1Id))
-
-        // syncing on this document with an unsupported spv should cause the document to desync
-        goOnline()
-        streamAndSync()
-
-        assertFalse(coll.syncedIds.contains(doc1Id))
-
-        // an error should also have been emitted
-        assertTrue(errorEmittedSem.tryAcquire(10, TimeUnit.SECONDS))
+    override fun testUnsupportedSpvFails() {
+        harness.testUnsupportedSpvFails()
     }
 
     @Test
-    fun testStaleFetchSingle() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val doc1 = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(doc1))
-
-            // get the document
-            val doc = Tasks.await(remoteColl.find(doc1).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                throw IllegalStateException("failure")
-            }, null, null)
-            coll.syncOne(doc1Id)
-
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("i", 1))))
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testStaleFetchSingle() {
+        harness.testStaleFetchSingle()
     }
 
     @Test
-    fun testStaleFetchSingleDeleted() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val doc1 = Document("hello", "world")
-            Tasks.await(remoteColl.insertOne(doc1))
-
-            // get the document
-            val doc = Tasks.await(remoteColl.find(doc1).first())!!
-            val doc1Id = BsonObjectId(doc.getObjectId("_id"))
-            val doc1Filter = Document("_id", doc1Id)
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                throw IllegalStateException("failure")
-            }, null, null)
-            coll.syncOne(doc1Id)
-
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("i", 1))))
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            assertEquals(1, Tasks.await(remoteColl.deleteOne(doc1Filter)).deletedCount)
-            powerCycleDevice()
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                throw IllegalStateException("failure")
-            }, null, null)
-
-            streamAndSync()
-            assertNull(Tasks.await(coll.findOneById(doc1Id)))
-        }
+    override fun testStaleFetchSingleDeleted() {
+        harness.testStaleFetchSingleDeleted()
     }
 
     @Test
-    fun testStaleFetchMultiple() {
-        testSyncInBothDirections {
-            val coll = getTestSync()
-
-            val remoteColl = getTestCollRemote()
-
-            val insertResult =
-                    Tasks.await(remoteColl.insertMany(listOf(
-                            Document("hello", "world"),
-                            Document("hello", "friend"))))
-
-            // get the document
-            val doc1Id = insertResult.insertedIds[0]
-            val doc2Id = insertResult.insertedIds[1]
-
-            coll.configure({ _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-                throw IllegalStateException("failure")
-            }, null, null)
-            coll.syncOne(doc1Id)
-
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            Tasks.await(coll.updateOneById(doc1Id, Document("\$inc", Document("i", 1))))
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-
-            coll.syncOne(doc2Id)
-            streamAndSync()
-            assertNotNull(Tasks.await(coll.findOneById(doc1Id)))
-            assertNotNull(Tasks.await(coll.findOneById(doc2Id)))
-        }
+    override fun testStaleFetchMultiple() {
+        harness.testStaleFetchMultiple()
     }
 
-    private fun streamAndSync() {
-        val dataSync = (mongoClient as RemoteMongoClientImpl).dataSynchronizer
-        if (testNetworkMonitor.connectedState) {
-            while (!dataSync.areAllStreamsOpen()) {
-                println("waiting for all streams to open before doing sync pass")
-                Thread.sleep(1000)
-            }
-        }
-        dataSync.doSyncPass()
-    }
-
-    private fun watchForEvents(
-        namespace: MongoNamespace,
-        n: Int = 1
-    ): Semaphore {
-        println("watching for $n change event(s) ns=$namespace")
-        val waitFor = AtomicInteger(n)
-        val sem = Semaphore(0)
-        (mongoClient as RemoteMongoClientImpl).dataSynchronizer.addWatcher(namespace, object : Callback<ChangeEvent<BsonDocument>, Any> {
-            override fun onComplete(result: OperationResult<ChangeEvent<BsonDocument>, Any>) {
-                if (result.isSuccessful && result.geResult() != null) {
-                    println("change event of operation ${result.geResult().operationType} ns=$namespace found!")
-                }
-                if (waitFor.decrementAndGet() == 0) {
-                    (mongoClient as RemoteMongoClientImpl).dataSynchronizer.removeWatcher(namespace, this)
-                    sem.release()
-                }
-            }
-        })
-        return sem
-    }
-
-    private fun powerCycleDevice() {
-        (mongoClient as RemoteMongoClientImpl).dataSynchronizer.reloadConfig()
-    }
-
-    private fun goOffline() {
-        println("going offline")
-        testNetworkMonitor.connectedState = false
-    }
-
-    private fun goOnline() {
-        println("going online")
-        testNetworkMonitor.connectedState = true
-    }
-
-    private fun withoutId(document: Document): Document {
-        val newDoc = Document(document)
-        newDoc.remove("_id")
-        return newDoc
-    }
-
-    private fun withoutSyncVersion(document: Document): Document {
-        val newDoc = Document(document)
-        newDoc.remove("__stitch_sync_version")
-        return newDoc
-    }
-
-    private fun withNewSyncVersionSet(document: Document): Document {
-        return appendDocumentToKey(
-                "\$set",
-                document,
-                Document("__stitch_sync_version", freshSyncVersionDoc()))
-    }
-
-    private fun withNewSyncVersion(document: Document): Document {
-        val newDocument = Document(java.util.HashMap(document))
-        newDocument["__stitch_sync_version"] = freshSyncVersionDoc()
-
-        return newDocument
-    }
-
-    private fun withNewUnsupportedSyncVersion(document: Document): Document {
-        val newDocument = Document(java.util.HashMap(document))
-        val badVersion = freshSyncVersionDoc()
-        badVersion.remove("spv")
-        badVersion.append("spv", 2)
-
-        newDocument["__stitch_sync_version"] = badVersion
-
-        return newDocument
-    }
-
-    private fun freshSyncVersionDoc(): Document {
-        return Document("spv", 1).append("id", UUID.randomUUID().toString()).append("v", 0L)
-    }
-
-    private fun versionOf(document: Document): Document {
-        return document["__stitch_sync_version"] as Document
-    }
-
-    private fun versionCounterOf(document: Document): Long {
-        return versionOf(document)["v"] as Long
-    }
-
-    private fun instanceIdOf(document: Document): String {
-        return versionOf(document)["id"] as String
-    }
-
-    private fun appendDocumentToKey(key: String, on: Document, toAppend: Document): Document {
-        val newDocument = Document(HashMap(on))
-        var found = false
-        newDocument.forEach {
-            if (it.key != key) {
-                return@forEach
-            }
-            found = true
-            val valueAtKey = (it.value as Document)
-            toAppend.forEach {
-                valueAtKey[it.key] = it.value
-            }
-        }
-        if (!found) {
-            newDocument[key] = toAppend
-        }
-        return newDocument
-    }
-
-    private val failingConflictHandler: ConflictHandler<Document> = ConflictHandler { _: BsonValue, _: ChangeEvent<Document>, _: ChangeEvent<Document> ->
-        fail("did not expect a conflict")
-        throw IllegalStateException("unreachable")
-    }
-
-    private fun testSyncInBothDirections(testFun: () -> Unit) {
-        val dataSync = (mongoClient as RemoteMongoClientImpl).dataSynchronizer
-        println("running tests with L2R going first")
-        dataSync.swapSyncDirection(true)
-        testFun()
-
-        teardown()
-        setup()
-        println("running tests with R2L going first")
-        dataSync.swapSyncDirection(false)
-        testFun()
+    /**
+     * Get the uri for where mongodb is running locally.
+     */
+    private fun getMongoDbUri(): String {
+        return InstrumentationRegistry.getArguments().getString(mongodbUriProp, "mongodb://localhost:26000")
     }
 }
