+++ core/testutils/build.gradle
@@ -9,6 +9,7 @@ buildscript {
 
 dependencies {
     implementation project(':core:stitch-core-sdk')
+    implementation project(':core:core-services:stitch-core-services-mongodb-remote')
     compile project(path: ':core:stitch-core-admin-client')
     api 'org.mongodb:bson:3.7.0'
     implementation 'junit:junit:4.12'
